+++ service/codestar/api.go
@@ -3,6 +3,7 @@
 package codestar
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -152,7 +153,10 @@ func (c *CodeStar) CreateProjectRequest(input *CreateProjectInput) (req *request
 
 // CreateProject API operation for AWS CodeStar.
 //
-// Reserved for future use. To create a project, use the AWS CodeStar console.
+// Creates a project, including project resources. This action creates a project
+// based on a submitted project request. A set of source code files and a toolchain
+// template file can be included with the project request. If these are not
+// provided, an empty project is created.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -1719,27 +1723,254 @@ func (s *AssociateTeamMemberOutput) SetClientRequestToken(v string) *AssociateTe
 	return s
 }
 
+// Location and destination information about the source code files provided
+// with the project request. The source code is uploaded to the new project
+// source repository after project creation.
+type Code struct {
+	_ struct{} `type:"structure"`
+
+	// The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit
+	// or GitHub. After AWS CodeStar provisions the new repository, the source code
+	// files provided with the project request are placed in the repository.
+	//
+	// Destination is a required field
+	Destination *CodeDestination `locationName:"destination" type:"structure" required:"true"`
+
+	// The location where the source code files provided with the project request
+	// are stored. AWS CodeStar retrieves the files during project creation.
+	//
+	// Source is a required field
+	Source *CodeSource `locationName:"source" type:"structure" required:"true"`
+}
+
+// String returns the string representation
+func (s Code) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s Code) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *Code) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "Code"}
+	if s.Destination == nil {
+		invalidParams.Add(request.NewErrParamRequired("Destination"))
+	}
+	if s.Source == nil {
+		invalidParams.Add(request.NewErrParamRequired("Source"))
+	}
+	if s.Destination != nil {
+		if err := s.Destination.Validate(); err != nil {
+			invalidParams.AddNested("Destination", err.(request.ErrInvalidParams))
+		}
+	}
+	if s.Source != nil {
+		if err := s.Source.Validate(); err != nil {
+			invalidParams.AddNested("Source", err.(request.ErrInvalidParams))
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetDestination sets the Destination field's value.
+func (s *Code) SetDestination(v *CodeDestination) *Code {
+	s.Destination = v
+	return s
+}
+
+// SetSource sets the Source field's value.
+func (s *Code) SetSource(v *CodeSource) *Code {
+	s.Source = v
+	return s
+}
+
+// Information about the AWS CodeCommit repository to be created in AWS CodeStar.
+// This is where the source code files provided with the project request will
+// be uploaded after project creation.
+type CodeCommitCodeDestination struct {
+	_ struct{} `type:"structure"`
+
+	// The name of the AWS CodeCommit repository to be created in AWS CodeStar.
+	//
+	// Name is a required field
+	Name *string `locationName:"name" min:"1" type:"string" required:"true"`
+}
+
+// String returns the string representation
+func (s CodeCommitCodeDestination) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s CodeCommitCodeDestination) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *CodeCommitCodeDestination) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "CodeCommitCodeDestination"}
+	if s.Name == nil {
+		invalidParams.Add(request.NewErrParamRequired("Name"))
+	}
+	if s.Name != nil && len(*s.Name) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("Name", 1))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetName sets the Name field's value.
+func (s *CodeCommitCodeDestination) SetName(v string) *CodeCommitCodeDestination {
+	s.Name = &v
+	return s
+}
+
+// The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit
+// or GitHub. After AWS CodeStar provisions the new repository, the source code
+// files provided with the project request are placed in the repository.
+type CodeDestination struct {
+	_ struct{} `type:"structure"`
+
+	// Information about the AWS CodeCommit repository to be created in AWS CodeStar.
+	// This is where the source code files provided with the project request will
+	// be uploaded after project creation.
+	CodeCommit *CodeCommitCodeDestination `locationName:"codeCommit" type:"structure"`
+
+	// Information about the GitHub repository to be created in AWS CodeStar. This
+	// is where the source code files provided with the project request will be
+	// uploaded after project creation.
+	GitHub *GitHubCodeDestination `locationName:"gitHub" type:"structure"`
+}
+
+// String returns the string representation
+func (s CodeDestination) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s CodeDestination) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *CodeDestination) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "CodeDestination"}
+	if s.CodeCommit != nil {
+		if err := s.CodeCommit.Validate(); err != nil {
+			invalidParams.AddNested("CodeCommit", err.(request.ErrInvalidParams))
+		}
+	}
+	if s.GitHub != nil {
+		if err := s.GitHub.Validate(); err != nil {
+			invalidParams.AddNested("GitHub", err.(request.ErrInvalidParams))
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetCodeCommit sets the CodeCommit field's value.
+func (s *CodeDestination) SetCodeCommit(v *CodeCommitCodeDestination) *CodeDestination {
+	s.CodeCommit = v
+	return s
+}
+
+// SetGitHub sets the GitHub field's value.
+func (s *CodeDestination) SetGitHub(v *GitHubCodeDestination) *CodeDestination {
+	s.GitHub = v
+	return s
+}
+
+// The location where the source code files provided with the project request
+// are stored. AWS CodeStar retrieves the files during project creation.
+type CodeSource struct {
+	_ struct{} `type:"structure"`
+
+	// Information about the Amazon S3 location where the source code files provided
+	// with the project request are stored.
+	//
+	// S3 is a required field
+	S3 *S3Location `locationName:"s3" type:"structure" required:"true"`
+}
+
+// String returns the string representation
+func (s CodeSource) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s CodeSource) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *CodeSource) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "CodeSource"}
+	if s.S3 == nil {
+		invalidParams.Add(request.NewErrParamRequired("S3"))
+	}
+	if s.S3 != nil {
+		if err := s.S3.Validate(); err != nil {
+			invalidParams.AddNested("S3", err.(request.ErrInvalidParams))
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetS3 sets the S3 field's value.
+func (s *CodeSource) SetS3(v *S3Location) *CodeSource {
+	s.S3 = v
+	return s
+}
+
 type CreateProjectInput struct {
 	_ struct{} `type:"structure"`
 
-	// Reserved for future use.
+	// A user- or system-generated token that identifies the entity that requested
+	// project creation. This token can be used to repeat the request.
 	ClientRequestToken *string `locationName:"clientRequestToken" min:"1" type:"string"`
 
-	// Reserved for future use.
+	// The description of the project, if any.
 	Description *string `locationName:"description" type:"string"`
 
-	// Reserved for future use.
+	// The ID of the project to be created in AWS CodeStar.
 	//
 	// Id is a required field
 	Id *string `locationName:"id" min:"2" type:"string" required:"true"`
 
-	// Reserved for future use.
+	// The display name for the project to be created in AWS CodeStar.
 	//
 	// Name is a required field
 	Name *string `locationName:"name" min:"1" type:"string" required:"true"`
 
+	// A list of the Code objects submitted with the project request. If this parameter
+	// is specified, the request must also include the toolchain parameter.
+	SourceCode []*Code `locationName:"sourceCode" type:"list"`
+
 	// The tags created for the project.
 	Tags map[string]*string `locationName:"tags" type:"map"`
+
+	// The name of the toolchain template file submitted with the project request.
+	// If this parameter is specified, the request must also include the sourceCode
+	// parameter.
+	Toolchain *Toolchain `locationName:"toolchain" type:"structure"`
 }
 
 // String returns the string representation
@@ -1769,6 +2000,21 @@ func (s *CreateProjectInput) Validate() error {
 	}
 	if s.Name != nil && len(*s.Name) < 1 {
 		invalidParams.Add(request.NewErrParamMinLen("Name", 1))
+	}
+	if s.SourceCode != nil {
+		for i, v := range s.SourceCode {
+			if v == nil {
+				continue
+			}
+			if err := v.Validate(); err != nil {
+				invalidParams.AddNested(fmt.Sprintf("%s[%v]", "SourceCode", i), err.(request.ErrInvalidParams))
+			}
+		}
+	}
+	if s.Toolchain != nil {
+		if err := s.Toolchain.Validate(); err != nil {
+			invalidParams.AddNested("Toolchain", err.(request.ErrInvalidParams))
+		}
 	}
 
 	if invalidParams.Len() > 0 {
@@ -1801,24 +2047,37 @@ func (s *CreateProjectInput) SetName(v string) *CreateProjectInput {
 	return s
 }
 
+// SetSourceCode sets the SourceCode field's value.
+func (s *CreateProjectInput) SetSourceCode(v []*Code) *CreateProjectInput {
+	s.SourceCode = v
+	return s
+}
+
 // SetTags sets the Tags field's value.
 func (s *CreateProjectInput) SetTags(v map[string]*string) *CreateProjectInput {
 	s.Tags = v
 	return s
 }
 
+// SetToolchain sets the Toolchain field's value.
+func (s *CreateProjectInput) SetToolchain(v *Toolchain) *CreateProjectInput {
+	s.Toolchain = v
+	return s
+}
+
 type CreateProjectOutput struct {
 	_ struct{} `type:"structure"`
 
-	// Reserved for future use.
+	// The Amazon Resource Name (ARN) of the created project.
 	//
 	// Arn is a required field
 	Arn *string `locationName:"arn" type:"string" required:"true"`
 
-	// Reserved for future use.
+	// A user- or system-generated token that identifies the entity that requested
+	// project creation.
 	ClientRequestToken *string `locationName:"clientRequestToken" min:"1" type:"string"`
 
-	// Reserved for future use.
+	// The ID of the project.
 	//
 	// Id is a required field
 	Id *string `locationName:"id" min:"2" type:"string" required:"true"`
@@ -2525,6 +2784,142 @@ func (s DisassociateTeamMemberOutput) GoString() string {
 	return s.String()
 }
 
+// Information about the GitHub repository to be created in AWS CodeStar. This
+// is where the source code files provided with the project request will be
+// uploaded after project creation.
+type GitHubCodeDestination struct {
+	_ struct{} `type:"structure"`
+
+	// Description for the GitHub repository to be created in AWS CodeStar. This
+	// description displays in GitHub after the repository is created.
+	Description *string `locationName:"description" min:"1" type:"string"`
+
+	// Whether to enable issues for the GitHub repository.
+	//
+	// IssuesEnabled is a required field
+	IssuesEnabled *bool `locationName:"issuesEnabled" type:"boolean" required:"true"`
+
+	// Name of the GitHub repository to be created in AWS CodeStar.
+	//
+	// Name is a required field
+	Name *string `locationName:"name" min:"1" type:"string" required:"true"`
+
+	// The GitHub username for the owner of the GitHub repository to be created
+	// in AWS CodeStar. If this repository should be owned by a GitHub organization,
+	// provide its name.
+	//
+	// Owner is a required field
+	Owner *string `locationName:"owner" min:"1" type:"string" required:"true"`
+
+	// Whether the GitHub repository is to be a private repository.
+	//
+	// PrivateRepository is a required field
+	PrivateRepository *bool `locationName:"privateRepository" type:"boolean" required:"true"`
+
+	// The GitHub user's personal access token for the GitHub repository.
+	//
+	// Token is a required field
+	Token *string `locationName:"token" min:"1" type:"string" required:"true"`
+
+	// The type of GitHub repository to be created in AWS CodeStar. Valid values
+	// are User or Organization.
+	//
+	// Type is a required field
+	Type *string `locationName:"type" type:"string" required:"true"`
+}
+
+// String returns the string representation
+func (s GitHubCodeDestination) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s GitHubCodeDestination) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *GitHubCodeDestination) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "GitHubCodeDestination"}
+	if s.Description != nil && len(*s.Description) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("Description", 1))
+	}
+	if s.IssuesEnabled == nil {
+		invalidParams.Add(request.NewErrParamRequired("IssuesEnabled"))
+	}
+	if s.Name == nil {
+		invalidParams.Add(request.NewErrParamRequired("Name"))
+	}
+	if s.Name != nil && len(*s.Name) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("Name", 1))
+	}
+	if s.Owner == nil {
+		invalidParams.Add(request.NewErrParamRequired("Owner"))
+	}
+	if s.Owner != nil && len(*s.Owner) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("Owner", 1))
+	}
+	if s.PrivateRepository == nil {
+		invalidParams.Add(request.NewErrParamRequired("PrivateRepository"))
+	}
+	if s.Token == nil {
+		invalidParams.Add(request.NewErrParamRequired("Token"))
+	}
+	if s.Token != nil && len(*s.Token) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("Token", 1))
+	}
+	if s.Type == nil {
+		invalidParams.Add(request.NewErrParamRequired("Type"))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetDescription sets the Description field's value.
+func (s *GitHubCodeDestination) SetDescription(v string) *GitHubCodeDestination {
+	s.Description = &v
+	return s
+}
+
+// SetIssuesEnabled sets the IssuesEnabled field's value.
+func (s *GitHubCodeDestination) SetIssuesEnabled(v bool) *GitHubCodeDestination {
+	s.IssuesEnabled = &v
+	return s
+}
+
+// SetName sets the Name field's value.
+func (s *GitHubCodeDestination) SetName(v string) *GitHubCodeDestination {
+	s.Name = &v
+	return s
+}
+
+// SetOwner sets the Owner field's value.
+func (s *GitHubCodeDestination) SetOwner(v string) *GitHubCodeDestination {
+	s.Owner = &v
+	return s
+}
+
+// SetPrivateRepository sets the PrivateRepository field's value.
+func (s *GitHubCodeDestination) SetPrivateRepository(v bool) *GitHubCodeDestination {
+	s.PrivateRepository = &v
+	return s
+}
+
+// SetToken sets the Token field's value.
+func (s *GitHubCodeDestination) SetToken(v string) *GitHubCodeDestination {
+	s.Token = &v
+	return s
+}
+
+// SetType sets the Type field's value.
+func (s *GitHubCodeDestination) SetType(v string) *GitHubCodeDestination {
+	s.Type = &v
+	return s
+}
+
 type ListProjectsInput struct {
 	_ struct{} `type:"structure"`
 
@@ -3084,6 +3479,55 @@ func (s *Resource) SetId(v string) *Resource {
 	return s
 }
 
+// The Amazon S3 location where the source code files provided with the project
+// request are stored.
+type S3Location struct {
+	_ struct{} `type:"structure"`
+
+	// The Amazon S3 object key where the source code files provided with the project
+	// request are stored.
+	BucketKey *string `locationName:"bucketKey" type:"string"`
+
+	// The Amazon S3 bucket name where the source code files provided with the project
+	// request are stored.
+	BucketName *string `locationName:"bucketName" min:"3" type:"string"`
+}
+
+// String returns the string representation
+func (s S3Location) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s S3Location) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *S3Location) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "S3Location"}
+	if s.BucketName != nil && len(*s.BucketName) < 3 {
+		invalidParams.Add(request.NewErrParamMinLen("BucketName", 3))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetBucketKey sets the BucketKey field's value.
+func (s *S3Location) SetBucketKey(v string) *S3Location {
+	s.BucketKey = &v
+	return s
+}
+
+// SetBucketName sets the BucketName field's value.
+func (s *S3Location) SetBucketName(v string) *S3Location {
+	s.BucketName = &v
+	return s
+}
+
 type TagProjectInput struct {
 	_ struct{} `type:"structure"`
 
@@ -3208,6 +3652,123 @@ func (s *TeamMember) SetRemoteAccessAllowed(v bool) *TeamMember {
 // SetUserArn sets the UserArn field's value.
 func (s *TeamMember) SetUserArn(v string) *TeamMember {
 	s.UserArn = &v
+	return s
+}
+
+// The toolchain template file provided with the project request. AWS CodeStar
+// uses the template to provision the toolchain stack in AWS CloudFormation.
+type Toolchain struct {
+	_ struct{} `type:"structure"`
+
+	// The service role ARN for AWS CodeStar to use for the toolchain template during
+	// stack provisioning.
+	RoleArn *string `locationName:"roleArn" min:"1" type:"string"`
+
+	// The Amazon S3 location where the toolchain template file provided with the
+	// project request is stored. AWS CodeStar retrieves the file during project
+	// creation.
+	//
+	// Source is a required field
+	Source *ToolchainSource `locationName:"source" type:"structure" required:"true"`
+
+	// The list of parameter overrides to be passed into the toolchain template
+	// during stack provisioning, if any.
+	StackParameters map[string]*string `locationName:"stackParameters" type:"map"`
+}
+
+// String returns the string representation
+func (s Toolchain) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s Toolchain) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *Toolchain) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "Toolchain"}
+	if s.RoleArn != nil && len(*s.RoleArn) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("RoleArn", 1))
+	}
+	if s.Source == nil {
+		invalidParams.Add(request.NewErrParamRequired("Source"))
+	}
+	if s.Source != nil {
+		if err := s.Source.Validate(); err != nil {
+			invalidParams.AddNested("Source", err.(request.ErrInvalidParams))
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetRoleArn sets the RoleArn field's value.
+func (s *Toolchain) SetRoleArn(v string) *Toolchain {
+	s.RoleArn = &v
+	return s
+}
+
+// SetSource sets the Source field's value.
+func (s *Toolchain) SetSource(v *ToolchainSource) *Toolchain {
+	s.Source = v
+	return s
+}
+
+// SetStackParameters sets the StackParameters field's value.
+func (s *Toolchain) SetStackParameters(v map[string]*string) *Toolchain {
+	s.StackParameters = v
+	return s
+}
+
+// The Amazon S3 location where the toolchain template file provided with the
+// project request is stored. AWS CodeStar retrieves the file during project
+// creation.
+type ToolchainSource struct {
+	_ struct{} `type:"structure"`
+
+	// The Amazon S3 bucket where the toolchain template file provided with the
+	// project request is stored.
+	//
+	// S3 is a required field
+	S3 *S3Location `locationName:"s3" type:"structure" required:"true"`
+}
+
+// String returns the string representation
+func (s ToolchainSource) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ToolchainSource) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *ToolchainSource) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "ToolchainSource"}
+	if s.S3 == nil {
+		invalidParams.Add(request.NewErrParamRequired("S3"))
+	}
+	if s.S3 != nil {
+		if err := s.S3.Validate(); err != nil {
+			invalidParams.AddNested("S3", err.(request.ErrInvalidParams))
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetS3 sets the S3 field's value.
+func (s *ToolchainSource) SetS3(v *S3Location) *ToolchainSource {
+	s.S3 = v
 	return s
 }